% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_server.R
\name{build_server}
\alias{build_server}
\title{Build the shiny server}
\usage{
build_server(wkdir)
}
\arguments{
\item{wkdir}{to Seurat RDS files to view}
}
\value{
shiny server
}
\description{
Build the shiny server
}
\keyword{internal}
