% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ui.R
\name{build_ui}
\alias{build_ui}
\title{Build the shiny server ui}
\usage{
build_ui(wkdir)
}
\arguments{
\item{wkdir}{to Seurat RDS files to view}
}
\value{
shiny ui
}
\description{
Build the shiny server ui
}
\keyword{internal}
