% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_summary}
\alias{data_summary}
\title{Creates a summarized data.frame aggregated by seurat_cluster}
\usage{
data_summary(query.df)
}
\arguments{
\item{query.df}{A data.frame}
}
\value{
data.frame
}
\description{
Creates a summarized data.frame aggregated by seurat_cluster
}
\keyword{internal}
