% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scatter.plot}
\alias{scatter.plot}
\title{Visualization code for scatter plot}
\usage{
scatter.plot(input_file)
}
\arguments{
\item{input_file}{A Seurat object with nFeature_RNA, nCount_RNA and seurat_clusters}
}
\value{
ggplot2 object
}
\description{
Visualization code for scatter plot
}
\keyword{internal}
