% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_main.R
\name{shinyQC}
\alias{shinyQC}
\title{Launch the shinyQC server}
\usage{
shinyQC(wkdir = ".", port = 5000)
}
\arguments{
\item{wkdir}{Path to Seurat RDS files to view}

\item{port}{Server port that will be exposed for connections}
}
\description{
This function initializes and launches a shiny server focused on \code{wkdir}. To connect locally
to a server running on a cluster environemnt use tunnelling:
}
\details{
\preformatted{ ssh -t AIBS.id@cluster.name -L 5000:localhost:5000
}
}
