% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{umap}
\alias{umap}
\title{Visualization code for umap plot}
\usage{
umap(
  input_file,
  selected = NULL,
  colour.by = "cluster",
  range.x = NULL,
  range.y = NULL
)
}
\arguments{
\item{input_file}{A Seurat object with UMAP dim. reduction and seurat_clusters}

\item{selected}{The cluster ids which have been selected via brush tool on scatter plot}

\item{colour.by}{Seurat meta.data field to colour the umap points}

\item{range.x}{x-axis limits for zooming}

\item{range.y}{y-axis limits for zooming}
}
\value{
ggplot2 object
}
\description{
Visualization code for umap plot
}
\keyword{internal}
